sarissa/requirements.txt
cached-property==1.5.2
Django==4.0.4
django-filter==21.1
django-url-filter==0.3.15
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
enum-compat==0.0.3
Markdown==3.3.7
psycopg2==2.9.3
PyJWT==2.3.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
